♻️ refactor: move components to upper folder and update import paths
