+++ src/app/page.tsx
@@ -2,8 +2,8 @@ import hero from '@/assets/images/hero.webp';
 import rain from '@/assets/images/rain.webp';
 import CommunityPop from '@/components/_home/CommunityPop';
 import { CurationCarousel } from '@/components/_home/CurationCarousel';
+import NewProductSection from '@/components/_home/NewProductSection';
 import ReviewCarousel from '@/components/_home/ReviewCarousel';
-import NewProductSection from '@/components/_home/shop-new/NewProductSection';
 import { Button } from '@/components/ui/Button';
 import Image from 'next/image';
 import Link from 'next/link';
